requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
wandb>=0.15.0
spacy>=3.6.0
nltk>=3.8
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
accelerate>=0.20.0
sentence-transformers>=2.2.0
networkx>=3.1
pydantic>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytest>=7.4.0
pytest-cov>=4.1.0 